# Documentation/openapi.yaml
openapi: 3.0.3
info:
  title: AIS Clinic API
  version: 1.0.0
  description: OpenAPI-описание ключевых эндпоинтов АИС "Клиника"
servers:
  - url: https://api.clinic.local
    description: Production
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/patients:
    post:
      tags: [Patients]
      summary: Регистрация пациента
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientCreateRequest'
      responses:
        '201':
          description: Пациент создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
    get:
      tags: [Patients]
      summary: Поиск пациентов
      parameters:
        - in: query
          name: q
          schema:
            type: string
            maxLength: 255
        - in: query
          name: birth_date
          schema:
            type: string
            format: date
        - in: query
          name: include_inactive
          schema:
            type: boolean
            default: false
        - in: query
          name: skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Список пациентов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/patients/{patient_id}:
    get:
      tags: [Patients]
      summary: Карточка пациента
      parameters:
        - in: path
          name: patient_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Пациент
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Patients]
      summary: Обновление данных пациента
      parameters:
        - in: path
          name: patient_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientUpdateRequest'
      responses:
        '200':
          description: Обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Patients]
      summary: Деактивация пациента (soft delete)
      parameters:
        - in: path
          name: patient_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Деактивирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/schedules:
    post:
      tags: [Schedules]
      summary: Создание расписания врача
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleCreateRequest'
      responses:
        '201':
          description: Расписание создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
    get:
      tags: [Schedules]
      summary: Список расписаний
      parameters:
        - in: query
          name: doctor_id
          schema:
            type: string
            format: uuid
        - in: query
          name: weekday
          schema:
            type: integer
            minimum: 1
            maximum: 7
        - in: query
          name: include_inactive
          schema:
            type: boolean
            default: false
        - in: query
          name: skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Список расписаний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/schedule-slots/generate:
    post:
      tags: [Schedules]
      summary: Генерация временных слотов по расписанию
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleSlotGenerateRequest'
      responses:
        '201':
          description: Слоты созданы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleSlotGenerateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/schedule-slots:
    get:
      tags: [Schedules]
      summary: Просмотр слотов по врачу и диапазону дат
      parameters:
        - in: query
          name: doctor_id
          schema:
            type: string
            format: uuid
        - in: query
          name: date_from
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: date_to
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: status
          schema:
            type: string
            enum: [available, booked, blocked, cancelled]
            default: available
        - in: query
          name: skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Список слотов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleSlotListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/schedule-exceptions:
    post:
      tags: [Schedules]
      summary: Создание исключения расписания (отпуск/больничный)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleExceptionCreateRequest'
      responses:
        '201':
          description: Исключение создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleExceptionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/appointments:
    post:
      tags: [Appointments]
      summary: Запись пациента на прием
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentCreateRequest'
      responses:
        '201':
          description: Запись создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
    get:
      tags: [Appointments]
      summary: Получить список записей
      parameters:
        - in: query
          name: doctor_id
          schema:
            type: string
            format: uuid
        - in: query
          name: patient_id
          schema:
            type: string
            format: uuid
        - in: query
          name: status
          schema:
            type: string
            enum: [planned, confirmed, cancelled, completed, no_show]
        - in: query
          name: skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Список записей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/reports/doctor-workload:
    get:
      tags: [Reports]
      summary: Отчет по загруженности врачей
      parameters:
        - in: query
          name: date_from
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: date_to
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Сводка по загруженности
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorWorkloadResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/reports/diagnoses:
    get:
      tags: [Reports]
      summary: Статистика диагнозов (ТОП)
      parameters:
        - in: query
          name: date_from
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: date_to
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
      responses:
        '200':
          description: Статистика диагнозов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnosesStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/reports/appointments:
    get:
      tags: [Reports]
      summary: Статистика записей за период
      parameters:
        - in: query
          name: date_from
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: date_to
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Сводка по статусам записей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentsStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/reports/finance:
    get:
      tags: [Reports]
      summary: Финансовая сводка по счетам за период
      parameters:
        - in: query
          name: date_from
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: date_to
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Финансовая статистика
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ValidationError:
      description: Ошибка валидации входных данных
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Требуется аутентификация
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Конфликт состояния
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    PatientCreateRequest:
      type: object
      required: [last_name, first_name, birth_date, gender]
      properties:
        last_name:
          type: string
          minLength: 1
          maxLength: 100
        first_name:
          type: string
          minLength: 1
          maxLength: 100
        middle_name:
          type: string
          maxLength: 100
        birth_date:
          type: string
          format: date
        gender:
          type: string
          enum: [M, F, O]
        phone:
          type: string
          pattern: '^\+?[0-9\-\(\)\s]{7,20}$'
        email:
          type: string
          format: email
        insurance_number:
          type: string
          maxLength: 50
        address:
          type: string
          maxLength: 500
    PatientResponse:
      type: object
      required: [id, last_name, first_name, birth_date, gender, is_active]
      properties:
        id:
          type: string
          format: uuid
        last_name:
          type: string
        first_name:
          type: string
        middle_name:
          type: string
          nullable: true
        birth_date:
          type: string
          format: date
        gender:
          type: string
          enum: [M, F, O]
        phone:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        insurance_number:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        is_active:
          type: boolean
    PatientListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PatientResponse'
        total:
          type: integer
          minimum: 0
    PatientUpdateRequest:
      type: object
      properties:
        last_name:
          type: string
          minLength: 1
          maxLength: 100
        first_name:
          type: string
          minLength: 1
          maxLength: 100
        middle_name:
          type: string
          maxLength: 100
        birth_date:
          type: string
          format: date
        gender:
          type: string
          enum: [M, F, O]
        phone:
          type: string
          pattern: '^\+?[0-9\-\(\)\s]{7,20}$'
        email:
          type: string
          format: email
        insurance_number:
          type: string
          maxLength: 50
        address:
          type: string
          maxLength: 500
    ScheduleCreateRequest:
      type: object
      required: [doctor_id, weekday, start_time, end_time, slot_minutes]
      properties:
        doctor_id:
          type: string
          format: uuid
        weekday:
          type: integer
          minimum: 1
          maximum: 7
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        slot_minutes:
          type: integer
          enum: [15, 30, 60]
        is_active:
          type: boolean
          default: true
    ScheduleResponse:
      type: object
      required: [id, doctor_id, weekday, start_time, end_time, slot_minutes, is_active]
      properties:
        id:
          type: string
          format: uuid
        doctor_id:
          type: string
          format: uuid
        weekday:
          type: integer
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        slot_minutes:
          type: integer
          enum: [15, 30, 60]
        is_active:
          type: boolean
    ScheduleListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleResponse'
        total:
          type: integer
          minimum: 0
    ScheduleSlotGenerateRequest:
      type: object
      required: [schedule_id, date_from, date_to]
      properties:
        schedule_id:
          type: string
          format: uuid
        date_from:
          type: string
          format: date
        date_to:
          type: string
          format: date
    ScheduleSlotResponse:
      type: object
      required: [id, schedule_id, start_at, end_at, status]
      properties:
        id:
          type: string
          format: uuid
        schedule_id:
          type: string
          format: uuid
        start_at:
          type: string
          format: date-time
        end_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [available, booked, blocked, cancelled]
    ScheduleSlotGenerateResponse:
      type: object
      required: [created, slots]
      properties:
        created:
          type: integer
          minimum: 0
        slots:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleSlotResponse'
    ScheduleSlotListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleSlotResponse'
        total:
          type: integer
          minimum: 0
    ScheduleExceptionCreateRequest:
      type: object
      required: [doctor_id, date_from, date_to, reason]
      properties:
        doctor_id:
          type: string
          format: uuid
        date_from:
          type: string
          format: date
        date_to:
          type: string
          format: date
        reason:
          type: string
          minLength: 3
          maxLength: 255
    ScheduleExceptionResponse:
      type: object
      required: [id, doctor_id, date_from, date_to, reason]
      properties:
        id:
          type: string
          format: uuid
        doctor_id:
          type: string
          format: uuid
        date_from:
          type: string
          format: date
        date_to:
          type: string
          format: date
        reason:
          type: string
    AppointmentCreateRequest:
      type: object
      required: [patient_id, doctor_id, slot_id]
      properties:
        patient_id:
          type: string
          format: uuid
        doctor_id:
          type: string
          format: uuid
        slot_id:
          type: string
          format: uuid
        reason:
          type: string
          maxLength: 2000
...
    AppointmentResponse:
      type: object
      required: [id, patient_id, doctor_id, slot_id, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        patient_full_name:
          type: string
          nullable: true
        doctor_id:
          type: string
          format: uuid
        doctor_full_name:
          type: string
          nullable: true
        slot_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [planned, confirmed, cancelled, completed, no_show]
        created_at:
          type: string
          format: date-time
    AppointmentListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AppointmentResponse'
        total:
          type: integer
          minimum: 0
...
    DoctorWorkloadItem:
      type: object
      required: [doctor_id, doctor_name, appointments_count, completed_count, cancelled_count]
      properties:
        doctor_id:
          type: string
          format: uuid
        doctor_name:
          type: string
        appointments_count:
          type: integer
          minimum: 0
        completed_count:
          type: integer
          minimum: 0
        cancelled_count:
          type: integer
          minimum: 0
    DoctorWorkloadResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DoctorWorkloadItem'
    DiagnosesStatsItem:
      type: object
      required: [icd10_code, title, total]
      properties:
        icd10_code:
          type: string
          pattern: '^[A-Z][0-9]{2}(\.[0-9A-Z]{1,2})?$'
        title:
          type: string
        total:
          type: integer
          minimum: 0
    DiagnosesStatsResponse:
      type: object
      required: [date_from, date_to, items]
      properties:
        date_from:
          type: string
          format: date
        date_to:
          type: string
          format: date
        items:
          type: array
          items:
            $ref: '#/components/schemas/DiagnosesStatsItem'
    AppointmentsStatsResponse:
      type: object
      required: [date_from, date_to, total, planned, confirmed, completed, cancelled, no_show]
      properties:
        date_from:
          type: string
          format: date
        date_to:
          type: string
          format: date
        total:
          type: integer
          minimum: 0
        planned:
          type: integer
          minimum: 0
        confirmed:
          type: integer
          minimum: 0
        completed:
          type: integer
          minimum: 0
        cancelled:
          type: integer
          minimum: 0
        no_show:
          type: integer
          minimum: 0
    FinanceStatsResponse:
      type: object
      required: [date_from, date_to, invoices_count, total_amount, paid_amount, issued_amount]
      properties:
        date_from:
          type: string
          format: date
        date_to:
          type: string
          format: date
        invoices_count:
          type: integer
          minimum: 0
        total_amount:
          type: number
          format: double
          minimum: 0
        paid_amount:
          type: number
          format: double
          minimum: 0
        issued_amount:
          type: number
          format: double
          minimum: 0
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  issue:
                    type: string
            request_id:
              type: string
              format: uuid